requirements.txt
Flask
loko-extensions
loko-client
loguru
langchain
openai
beautifulsoup4
tiktoken
faiss-cpu
